added tests for waitlist
